added new models: Alex, Dop, Dopoln, Person, Persons, Prov, Robot, Robots
