+++ Qwe/Models/Dopoln.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace Qwe.Models
+{
+    public class Dopoln
+    {
+        protected internal bool Hello { get; set; }
+        protected private int Sum { get; set; }
+    }
+}
