+++ Qwe/Models/Robot.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace Qwe.Models
+{
+    public class Robot
+    {
+        static Robot() { }
+    }
+}
